refactor : 권한 요청 메소드 분리
